backend/requirements.txt
flask
pymupdf
Pillow
pytesseract
gTTS
python-dotenv
Flask-Cors
